% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRVmat.R
\name{computeRVmat}
\alias{computeRVmat}
\title{Compute a similarity or dissimilarity matrix with all pairwise modified RV-coefficients}
\usage{
computeRVmat(DataList = DataList, dist = TRUE, verbose = TRUE)
}
\arguments{
\item{DataList}{a list with N matrices}

\item{dist}{boolean, if TRUE a distance matrix is return, if FALSE a similarity matrix is return}

\item{verbose}{if TRUE a progresbar is printed to the console}
}
\value{
a NxN similarity matrix or a dist object (dependening on the dist argument)
}
\description{
This function computes all pairwise modified RV-coefficients between the matrices that are stored in the DataList object
}
\examples{
Mat <- matrix(rnorm(1000), 100)
SVD <- svd(Mat)
X <- vector(mode = 'list', length = 10)
for(i in 1:5){
X[[i]] <- SVD$u[,-i] \%*\% diag(SVD$d[-i]) \%*\% t(SVD$v[,-i])
}
X[[6]] <- matrix(rnorm(1000, mean = 0), 100)
SVD <- svd(X[[6]])
for(i in 6:10){
X[[i]] <- SVD$u[,-(i-5)] \%*\% diag(SVD$d[-(i-5)]) \%*\% t(SVD$v[,-(i-5)])
}

RVmat <- computeRVmat(DataList = X, dist = FALSE, verbose = TRUE)
heatmap(RVmat)
}
\seealso{
\code{\link{modRV}, \link{icaList}}
}
