% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icaList.R
\name{icaList}
\alias{icaList}
\title{icaList, wrapper function to do single subject ICA analyses on a list of matrices}
\usage{
icaList(DataList, nc, outputS = TRUE, verbose = TRUE)
}
\arguments{
\item{DataList, }{list with datasets. Each matrix is a dataset to do ICA on}

\item{nc}{number of components to select for ICA}

\item{outputS}{boolean, if TRUE only the ICA components are returned. If FALSE, all icafast output is returned}

\item{verbose}{boolean,if TRUE a progressbar is printed to the screen}
}
\value{
a list where each element is a matrix with single subject independent components
}
\description{
icaList, wrapper function to do single subject ICA analyses on a list of matrices
}
\examples{
set.seed(42)
S <- vector('list', length = 4)
X <- vector('list', length = 4)
for(i in 1:4){
S[[i]] <- cbind(runif(1000), runif(1000))
A <- matrix(data = runif(4, -2, 2), nrow = 2)
X[[i]] <- S[[i]] \%*\% t(A)
}
Output <- icaList(X, 2, outputS = T, verbose = T)
lapply(seq_along(Output), function(lambda) cor(Output[[lambda]], S[[lambda]] ) )
}
\seealso{
\code{\link{computeRVmat}}
}
